Fleshed out the collision response and added jumping
The collision response has gotten pretty complicated now. It will push
the player the shortest possible way out of a collider2d (axis aligned,
so only along x or y axis) and repeat for all things it is colliding
with a bunch of times each frame before giving up on finding a solution
to where the player should be. Jumping is implemented as moving upward
at a fixed speed for the duration of your jump time. Jump time is
limited, and can be cut short by releasing the control early. Separated
max speed for horizontal and vertical axes. Also changed to text
serialization so merge conflicts aren't completely possible to fix.